% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegCIT168-package.R
\docType{package}
\name{ggsegCIT168-package}
\alias{ggsegCIT168}
\alias{ggsegCIT168-package}
\title{ggsegCIT168: CIT168 Subcortical Atlas for the 'ggsegverse' Ecosystem}
\description{
CIT168 probabilistic subcortical atlas (Pauli et al. 2018) for the 'ggseg' ecosystem. Provides a 'ggseg_atlas' object with 2D polygon geometry and 3D mesh data, covering the striatum, pallidum and the small midbrain and diencephalic nuclei, for use with 'ggseg' and 'ggseg3d'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegCIT168}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegCIT168/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})
  \item Didac Vidal-Pineiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

Other contributors:
\itemize{
  \item Center for Lifespan Changes in Brain and Cognition (LCBC), University of Oslo [copyright holder]
}

}
\keyword{internal}
